Uses sample data for weight entries
Populates the weight entry list with a short sample data set.

This allows for testing and UI development without relying on Core Data.
It also ensures a default state for new users.

Hides the query to the Core Data.

If goal weight is set to zero, then the chart doesn't display the Goal Weight line.